Move posthog to a deferred load to reduce TTL
