+++ package.json
@@ -19,7 +19,8 @@
     "test-e2e-ui": "npx playwright test --ui",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
-    "analyze": "THEME=cpr next experimental-analyze"
+    "analyze": "THEME=cpr next experimental-analyze",
+    "size": "npx size-limit"
   },
   "dependencies": {
     "@base-ui/react": "^1.5.0",
