+++ src/utils/features.ts
@@ -1,5 +1,3 @@
-import { useFeatureFlagEnabled } from "posthog-js/react";
-
 import { AB_TEST_KEYS, CONFIG_FEATURE_KEYS, FEATURE_FLAG_KEYS, TConfigFeature, TFeatureFlag, TFeatureFlags, TFeatures, TThemeConfig } from "@/types";
 
 interface IArgs {
@@ -39,16 +37,4 @@ export const getFeatures = (themeConfig: TThemeConfig, featureFlags: TFeatureFla
   }
 
   return features;
-};
-
-// Add A/B test feature flags into TFeatures. Does not work if called in getServerSideProps
-export const getFeaturesWithABTests = (features: TFeatures): TFeatures => {
-  const featuresWithABTests = { ...features };
-
-  AB_TEST_KEYS.forEach((key) => {
-    const abTestIsEnabled = useFeatureFlagEnabled(key);
-    featuresWithABTests[key] = Boolean(abTestIsEnabled); // May be undefined if not yet initialised
-  });
-
-  return featuresWithABTests;
 };
